+++ generators/constant/inputs/edit.js
@@ -3,20 +3,20 @@
  * Generated file edit on WebStorm
  *
  * Created by: Geoffrey "C0ZEN" Testelin
- * Date: 07/01/2018
- * Time: 19:08
- * Version: 0.12.5
+ * Date: 08/01/2018
+ * Time: 00:09
+ * Version: 0.12.6
  */
 (function () {
 	'use strict';
 
-	const logs           = require('../../common/logs.js');
-	const q              = require('q');
-	const controllerName = require('./controllerName.js');
-	const controllerPath = require('./controllerPath.js');
-	const fileName       = require('./fileName.js');
+	const logs         = require('../../common/logs.js');
+	const q            = require('q');
+	const constantName = require('./constantName.js');
+	const constantPath = require('./constantPath.js');
+	const fileName     = require('./fileName.js');
 
-	const longerGenerator = 15;
+	const longerGenerator = 13;
 	const prompts         = [
 		{
 			type   : 'list',
@@ -24,8 +24,8 @@
 			message: 'Choose a value to edit:',
 			choices: [
 				{
-					value  : 'controllerName',
-					short  : 'controllerName',
+					value  : 'constantName',
+					short  : 'constantName',
 					checked: true
 				},
 				{
@@ -33,8 +33,8 @@
 					short: 'fileName'
 				},
 				{
-					value: 'controllerPath',
-					short: 'controllerPath'
+					value: 'constantPath',
+					short: 'constantPath'
 				}
 			]
 		}
@@ -51,9 +51,9 @@
 			logs.hintsMultiline($that, hints);
 
 			/* eslint no-magic-numbers:"off" */
-			prompts[0].choices[0].name = logs.choice('Controller name', $that.controllerName + 'Controller', longerGenerator);
+			prompts[0].choices[0].name = logs.choice('Constant name', $that.constantName + 'Constant', longerGenerator);
 			prompts[0].choices[1].name = logs.choice('File name', $that.fileName, longerGenerator);
-			prompts[0].choices[2].name = logs.choice('File path', $that.controllerPath, longerGenerator);
+			prompts[0].choices[2].name = logs.choice('File path', $that.constantPath, longerGenerator);
 
 			$that.prompt(prompts).then($response => {
 				onSuccess($response);
@@ -65,14 +65,14 @@
 				$that.log();
 
 				switch ($response.edit) {
-					case 'controllerName':
-						deferred.resolve(controllerName.set($that, true));
+					case 'constantName':
+						deferred.resolve(constantName.set($that, true));
 						break;
 					case 'fileName':
 						deferred.resolve(fileName.set($that, true));
 						break;
-					case 'controllerPath':
-						deferred.resolve(controllerPath.set($that, true));
+					case 'constantPath':
+						deferred.resolve(constantPath.set($that, true));
 						break;
 					default:
 						break;